Patch on integer overflow for stata limits that are below the actual limit of the int type

+++ src/mapping.rs
@@ -31,8 +31,8 @@ pub enum StataType {
 impl StataType {
     pub fn to_string(&self) -> &'static str {
         match self {
-            StataType::Byte => "byte",
-            StataType::Int => "int",
+            StataType::Byte => "int",
+            StataType::Int => "long",
             StataType::Long => "long",
             StataType::Float => "float",
             StataType::Double => "double",
@@ -56,11 +56,11 @@ pub fn map_polars_to_stata(
         //  Boolean
         DataType::Boolean => StataType::Byte,
         // Integers
-        DataType::Int8 => StataType::Byte,
-        DataType::Int16 => StataType::Int,
+        DataType::Int8 => StataType::Int,
+        DataType::Int16 => StataType::Long,
         DataType::Int32 => StataType::Long,
         DataType::Int64 => StataType::Double, // Only double can contain the set of possible values
-        DataType::UInt8 => StataType::Byte, 
+        DataType::UInt8 => StataType::Int, 
         DataType::UInt16 => StataType::Long,
         DataType::UInt32 => StataType::Long,
         DataType::UInt64 => StataType::Double,
